+++ .travis.yml
@@ -25,102 +25,102 @@ matrix:
       env: TOX_POSARGS="-e py35-installation"
     # geth 1.5.6
     - python: "2.7"
-      env: TOX_POSARGS="-e py27-stdlib -e py27-gevent" GETH_VERSION=v1.5.6
+      env: TOX_POSARGS="-e py27-core" GETH_VERSION=v1.5.6
     - python: "3.4"
-      env: TOX_POSARGS="-e py34-stdlib -e py34-gevent" GETH_VERSION=v1.5.6
+      env: TOX_POSARGS="-e py34-core" GETH_VERSION=v1.5.6
     - python: "3.5"
-      env: TOX_POSARGS="-e py35-stdlib -e py35-gevent" GETH_VERSION=v1.5.6
+      env: TOX_POSARGS="-e py35-core" GETH_VERSION=v1.5.6
     # geth 1.5.7
     - python: "2.7"
-      env: TOX_POSARGS="-e py27-stdlib -e py27-gevent" GETH_VERSION=v1.5.7
+      env: TOX_POSARGS="-e py27-core" GETH_VERSION=v1.5.7
     - python: "3.4"
-      env: TOX_POSARGS="-e py34-stdlib -e py34-gevent" GETH_VERSION=v1.5.7
+      env: TOX_POSARGS="-e py34-core" GETH_VERSION=v1.5.7
     - python: "3.5"
-      env: TOX_POSARGS="-e py35-stdlib -e py35-gevent" GETH_VERSION=v1.5.7
+      env: TOX_POSARGS="-e py35-core" GETH_VERSION=v1.5.7
     # geth 1.5.8
     - python: "2.7"
-      env: TOX_POSARGS="-e py27-stdlib -e py27-gevent" GETH_VERSION=v1.5.8
+      env: TOX_POSARGS="-e py27-core" GETH_VERSION=v1.5.8
     - python: "3.4"
-      env: TOX_POSARGS="-e py34-stdlib -e py34-gevent" GETH_VERSION=v1.5.8
+      env: TOX_POSARGS="-e py34-core" GETH_VERSION=v1.5.8
     - python: "3.5"
-      env: TOX_POSARGS="-e py35-stdlib -e py35-gevent" GETH_VERSION=v1.5.8
+      env: TOX_POSARGS="-e py35-core" GETH_VERSION=v1.5.8
     # geth 1.5.9
     - python: "2.7"
-      env: TOX_POSARGS="-e py27-stdlib -e py27-gevent" GETH_VERSION=v1.5.9
+      env: TOX_POSARGS="-e py27-core" GETH_VERSION=v1.5.9
     - python: "3.4"
-      env: TOX_POSARGS="-e py34-stdlib -e py34-gevent" GETH_VERSION=v1.5.9
+      env: TOX_POSARGS="-e py34-core" GETH_VERSION=v1.5.9
     - python: "3.5"
-      env: TOX_POSARGS="-e py35-stdlib -e py35-gevent" GETH_VERSION=v1.5.9
+      env: TOX_POSARGS="-e py35-core" GETH_VERSION=v1.5.9
     # geth 1.6.0
     - python: "2.7"
-      env: TOX_POSARGS="-e py27-stdlib -e py27-gevent" GETH_VERSION=v1.6.0
+      env: TOX_POSARGS="-e py27-core" GETH_VERSION=v1.6.0
     - python: "3.4"
-      env: TOX_POSARGS="-e py34-stdlib -e py34-gevent" GETH_VERSION=v1.6.0
+      env: TOX_POSARGS="-e py34-core" GETH_VERSION=v1.6.0
     - python: "3.5"
-      env: TOX_POSARGS="-e py35-stdlib -e py35-gevent" GETH_VERSION=v1.6.0
+      env: TOX_POSARGS="-e py35-core" GETH_VERSION=v1.6.0
     # geth 1.6.1
     - python: "2.7"
-      env: TOX_POSARGS="-e py27-stdlib -e py27-gevent" GETH_VERSION=v1.6.1
+      env: TOX_POSARGS="-e py27-core" GETH_VERSION=v1.6.1
     - python: "3.4"
-      env: TOX_POSARGS="-e py34-stdlib -e py34-gevent" GETH_VERSION=v1.6.1
+      env: TOX_POSARGS="-e py34-core" GETH_VERSION=v1.6.1
     - python: "3.5"
-      env: TOX_POSARGS="-e py35-stdlib -e py35-gevent" GETH_VERSION=v1.6.1
+      env: TOX_POSARGS="-e py35-core" GETH_VERSION=v1.6.1
     # geth 1.6.2
     - python: "2.7"
-      env: TOX_POSARGS="-e py27-stdlib -e py27-gevent" GETH_VERSION=v1.6.2
+      env: TOX_POSARGS="-e py27-core" GETH_VERSION=v1.6.2
     - python: "3.4"
-      env: TOX_POSARGS="-e py34-stdlib -e py34-gevent" GETH_VERSION=v1.6.2
+      env: TOX_POSARGS="-e py34-core" GETH_VERSION=v1.6.2
     - python: "3.5"
-      env: TOX_POSARGS="-e py35-stdlib -e py35-gevent" GETH_VERSION=v1.6.2
+      env: TOX_POSARGS="-e py35-core" GETH_VERSION=v1.6.2
     # geth 1.6.3
     - python: "2.7"
-      env: TOX_POSARGS="-e py27-stdlib -e py27-gevent" GETH_VERSION=v1.6.3
+      env: TOX_POSARGS="-e py27-core" GETH_VERSION=v1.6.3
     - python: "3.4"
-      env: TOX_POSARGS="-e py34-stdlib -e py34-gevent" GETH_VERSION=v1.6.3
+      env: TOX_POSARGS="-e py34-core" GETH_VERSION=v1.6.3
     - python: "3.5"
-      env: TOX_POSARGS="-e py35-stdlib -e py35-gevent" GETH_VERSION=v1.6.3
+      env: TOX_POSARGS="-e py35-core" GETH_VERSION=v1.6.3
     # geth 1.6.4
     - python: "2.7"
-      env: TOX_POSARGS="-e py27-stdlib -e py27-gevent" GETH_VERSION=v1.6.4
+      env: TOX_POSARGS="-e py27-core" GETH_VERSION=v1.6.4
     - python: "3.4"
-      env: TOX_POSARGS="-e py34-stdlib -e py34-gevent" GETH_VERSION=v1.6.4
+      env: TOX_POSARGS="-e py34-core" GETH_VERSION=v1.6.4
     - python: "3.5"
-      env: TOX_POSARGS="-e py35-stdlib -e py35-gevent" GETH_VERSION=v1.6.4
+      env: TOX_POSARGS="-e py35-core" GETH_VERSION=v1.6.4
     # geth 1.6.5
     - python: "2.7"
-      env: TOX_POSARGS="-e py27-stdlib -e py27-gevent" GETH_VERSION=v1.6.5
+      env: TOX_POSARGS="-e py27-core" GETH_VERSION=v1.6.5
     - python: "3.4"
-      env: TOX_POSARGS="-e py34-stdlib -e py34-gevent" GETH_VERSION=v1.6.5
+      env: TOX_POSARGS="-e py34-core" GETH_VERSION=v1.6.5
     - python: "3.5"
-      env: TOX_POSARGS="-e py35-stdlib -e py35-gevent" GETH_VERSION=v1.6.5
+      env: TOX_POSARGS="-e py35-core" GETH_VERSION=v1.6.5
     # geth 1.6.6
     - python: "2.7"
-      env: TOX_POSARGS="-e py27-stdlib -e py27-gevent" GETH_VERSION=v1.6.6
+      env: TOX_POSARGS="-e py27-core" GETH_VERSION=v1.6.6
     - python: "3.4"
-      env: TOX_POSARGS="-e py34-stdlib -e py34-gevent" GETH_VERSION=v1.6.6
+      env: TOX_POSARGS="-e py34-core" GETH_VERSION=v1.6.6
     - python: "3.5"
-      env: TOX_POSARGS="-e py35-stdlib -e py35-gevent" GETH_VERSION=v1.6.6
+      env: TOX_POSARGS="-e py35-core" GETH_VERSION=v1.6.6
     # geth 1.6.7
     - python: "2.7"
-      env: TOX_POSARGS="-e py27-stdlib -e py27-gevent" GETH_VERSION=v1.6.7
+      env: TOX_POSARGS="-e py27-core" GETH_VERSION=v1.6.7
     - python: "3.4"
-      env: TOX_POSARGS="-e py34-stdlib -e py34-gevent" GETH_VERSION=v1.6.7
+      env: TOX_POSARGS="-e py34-core" GETH_VERSION=v1.6.7
     - python: "3.5"
-      env: TOX_POSARGS="-e py35-stdlib -e py35-gevent" GETH_VERSION=v1.6.7
+      env: TOX_POSARGS="-e py35-core" GETH_VERSION=v1.6.7
     # geth 1.7.0
     - python: "2.7"
-      env: TOX_POSARGS="-e py27-stdlib -e py27-gevent" GETH_VERSION=v1.7.0
+      env: TOX_POSARGS="-e py27-core" GETH_VERSION=v1.7.0
     - python: "3.4"
-      env: TOX_POSARGS="-e py34-stdlib -e py34-gevent" GETH_VERSION=v1.7.0
+      env: TOX_POSARGS="-e py34-core" GETH_VERSION=v1.7.0
     - python: "3.5"
-      env: TOX_POSARGS="-e py35-stdlib -e py35-gevent" GETH_VERSION=v1.7.0
+      env: TOX_POSARGS="-e py35-core" GETH_VERSION=v1.7.0
     # geth 1.7.2
     - python: "2.7"
-      env: TOX_POSARGS="-e py27-stdlib -e py27-gevent" GETH_VERSION=v1.7.2
+      env: TOX_POSARGS="-e py27-core" GETH_VERSION=v1.7.2
     - python: "3.4"
-      env: TOX_POSARGS="-e py34-stdlib -e py34-gevent" GETH_VERSION=v1.7.2
+      env: TOX_POSARGS="-e py34-core" GETH_VERSION=v1.7.2
     - python: "3.5"
-      env: TOX_POSARGS="-e py35-stdlib -e py35-gevent" GETH_VERSION=v1.7.2
+      env: TOX_POSARGS="-e py35-core" GETH_VERSION=v1.7.2
 before_install:
   - travis_retry sudo add-apt-repository -y ppa:gophers/archive
   - travis_retry sudo apt-get update
