+++ geth/utils/thread.py
@@ -1,0 +1,12 @@
+import threading
+
+
+def spawn(target, *args, **kwargs):
+    thread = threading.Thread(
+        target=target,
+        args=args,
+        kwargs=kwargs,
+    )
+    thread.daemon = True
+    thread.start()
+    return thread
