+++ geth/utils/timeout.py
@@ -1,20 +1,4 @@
-"""
-A minimal implementation of the various gevent APIs used within this codebase.
-"""
-import sys
 import time
-import threading
-import socket  # noqa: F401
-import subprocess  # noqa: F401
-from wsgiref.simple_server import make_server  # noqa: F401
-
-if sys.version_info.major == 2:
-    import Queue as queue
-else:
-    import queue
-
-
-sleep = time.sleep
 
 
 class Timeout(Exception):
@@ -74,53 +58,3 @@ class Timeout(Exception):
 
     def cancel(self):
         self.is_running = False
-
-
-class empty(object):
-    pass
-
-
-class ThreadWithReturn(threading.Thread):
-    def __init__(self, target=None, args=None, kwargs=None):
-        super(ThreadWithReturn, self).__init__(target=target, args=args, kwargs=kwargs)
-        self.target = target
-        self.args = args
-        self.kwargs = kwargs
-
-    def run(self):
-        self._return = self.target(*self.args, **self.kwargs)
-
-    def get(self, timeout=None):
-        self.join(timeout)
-        try:
-            return self._return
-        except AttributeError:
-            raise RuntimeError("Something went wrong.  No `_return` property was set")
-
-
-def spawn(target, *args, **kwargs):
-    thread = ThreadWithReturn(
-        target=target,
-        args=args,
-        kwargs=kwargs,
-    )
-    thread.daemon = True
-    thread.start()
-    return thread
-
-
-class JoinableQueue(queue.Queue):
-    def __iter__(self):
-        while True:
-            item = self.get()
-            if isinstance(item, Exception):
-                raise item
-            elif isinstance(item, type) and issubclass(item, Exception):
-                raise item
-            yield item
-
-    def join(self, timeout=None):
-        with Timeout(timeout) as _timeout:
-            while not self.empty():
-                sleep(0)
-                _timeout.check()
