--- requirements-gevent.txt
@@ -1,3 +1,0 @@
--r requirements-dev.txt
-gevent>=1.1.1,<1.2
-geventhttpclient>=1.3.1
