--- requirements.txt
@@ -1,1 +1,0 @@
-gevent>=1.1.2
