+++ tests/core/running/test_running_with_logging.py
