+++ tests/core/utility/test_is_live_chain.py
