+++ tox.ini
@@ -1,7 +1,6 @@
 [tox]
 envlist=
-    py{27,34,35}-{gevent,stdlib},
-    py{27,34,35}-installation
+    py{27,34,35}-{core,installation},
     flake8
 
 [flake8]
@@ -10,8 +9,8 @@ exclude= tests/*
 
 [testenv]
 commands=
-    {gevent,stdlib}: py.test {posargs:tests}
-    py{27,34,35}-installation: py.test {posargs:-s tests/installation}
+    core: py.test {posargs:tests/core}
+    installation: py.test {posargs:-s tests/installation}
 passenv =
     GETH_BASE_INSTALL_PATH
     GETH_BINARY
@@ -19,11 +18,9 @@ passenv =
     GOROOT
     GOPATH
 setenv =
-    gevent: GETH_THREADING_BACKEND=gevent
-    py{27,34,35}-installation: GETH_RUN_INSTALL_TESTS=enabled
+    installation: GETH_RUN_INSTALL_TESTS=enabled
 deps =
     -r{toxinidir}/requirements-dev.txt
-    gevent: -r{toxinidir}/requirements-gevent.txt
 basepython =
     py27: python2.7
     py34: python3.4
